# java: Improve bnechmark using BufferedReader