Here is a significant improvement and competitors should use `BufferedReader` to avoid unwanted TLE.

Most competitors now use `Scanner` to utilize `nextInt` or so. But this manner is known slow. With Procon Input, competitors can parse the input faster than ever.

before (`Scanner#nextLine`):
```
java-1: 225.300588ms
java-2: 366.386464ms
java-3: 145.577173ms
```

after (`BufferedReader#readLine`):
```
java-1: 142.859143ms
java-2: 166.40148ms
java-3: 103.147926ms
```