@akiradeveloper 2023-01-19
Here is the bench of all languages. Umm, I think C++ can improve.

```
cpp-1: 63.868056ms
cpp-2: 208.080363ms
cpp-3: 33.688905ms
csharp-1: 72.942699ms
/tmp/Main.cs(9,5): warning CS0219: The variable `n' is assigned but its value is never used
Compilation succeeded - 1 warning(s)
csharp-2: 123.523028ms
csharp-3: 43.723153ms
java-1: 123.703993ms
java-2: 158.194948ms
java-3: 111.496991ms
CC: stdlib_io.nim
CC: stdlib_system.nim
CC: stdlib_parseutils.nim
CC: stdlib_strutils.nim
CC: stdlib_sequtils.nim
CC: main.nim
nim-1: 161.277279ms
CC: stdlib_system.nim
CC: stdlib_parseutils.nim
CC: stdlib_strutils.nim
CC: main.nim
nim-2: 613.609634ms
CC: stdlib_system.nim
CC: main.nim
nim-3: 45.491702ms
python-1: 53.225014ms
python-2: 141.114251ms
python-3: 25.149306ms
ruby-1: 86.001063ms
ruby-2: 193.68753ms
ruby-3: 61.553792ms
rust-1: 22.556533ms
rust-2: 111.930661ms
rust-3: 9.454423ms
```